html/infrastructure/safe-passing-of-structured-data/shared-array-buffers/identity-not-preserved.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>SharedArrayBuffers, when cloned, do not give back the same object</title>
<link rel="help" href="https://html.spec.whatwg.org/multipage/#structuredserialize">
<link rel="author" title="Domenic Denicola" href="mailto:d@domenic.me">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<div id="log"></div>

<script>
"use strict";

async_test(t => {
  const sab = new SharedArrayBuffer(1);
  window.addEventListener("message", t.step_func(({ data }) => {
    if (data.testId !== 1) {
      return;
    }

    assert_not_equals(data.sab, sab);

    const originalView = new Uint8Array(sab);
    const receivedView = new Uint8Array(data.sab);
    assert_equals(originalView[0], 0, "originalView[0] starts 0");
    assert_equals(receivedView[0], 0, "receivedView[0] starts 0");

    originalView[0] = 5;
    assert_equals(originalView[0], 5, "originalView[0] ends up 5");
    assert_equals(receivedView[0], 5, "receivedView[0] ends up 5");

    t.done();
  }));

  window.postMessage({ sab, testId: 1 }, "*");
}, "postMessaging to this window does not give back the same SharedArrayBuffer (but does use the same backing block)");

async_test(t => {
  const sab = new SharedArrayBuffer();
  const worker = new Worker("resources/echo-worker.js");

  worker.addEventListener("message", t.step_func(({ data }) => {
    if (data.testId !== 2) {
      return;
    }

    assert_not_equals(data.sab, sab);
    t.done();
  }));

  worker.postMessage({ testId: 2, sab });
}, "postMessaging to a worker and back does not give back the same SharedArrayBuffer");

async_test(t => {
  const sab = new SharedArrayBuffer();
  window.addEventListener("message", t.step_func(({ data }) => {
    if (data.testId !== 3) {
      return;
    }

    assert_not_equals(data.sab, sab);
    t.done();
  }));

  const iframe = document.createElement("iframe");
  iframe.onload = t.step_func(() => {
    iframe.contentWindow.postMessage({ testId: 3, sab }, "*");
  });
  iframe.src = "resources/echo-iframe.html";
  document.body.appendChild(iframe);
}, "postMessaging to an iframe and back does not give back the same SharedArrayBuffer");
</script>
